Address delete/update/read operations

+++ src/main/java/com/qrzlo/webshop/data/repository/AddressRepository.java
@@ -6,7 +6,7 @@ import org.springframework.data.repository.CrudRepository;
 
 import java.util.List;
 
-public interface AddressRepository extends CrudRepository<Address, Long>
+public interface AddressRepository extends CrudRepository<Address, Integer>
 {
 	List<Address> findAddressesByCustomer(Customer customer);
 }
